Fix compilation errors with gcc.

+++ TextDrawer.cpp
@@ -107,7 +107,7 @@ struct Atlas {
 				rowh = 0;
 			}
 			roww += g->bitmap.width + 1;
-			rowh = std::max(rowh, g->bitmap.rows);
+			rowh = std::max(rowh, (int)g->bitmap.rows);
 		}
 
 		w = std::max(w, roww);
@@ -162,7 +162,7 @@ struct Atlas {
 			c[i].tx = ox / (float)w;
 			c[i].ty = oy / (float)h;
 
-			rowh = std::max(rowh, g->bitmap.rows);
+			rowh = std::max(rowh, (int)g->bitmap.rows);
 			ox += g->bitmap.width + 1;
 		}
